Make UART functions borrow self mutably.
This isn't enforced by the `earlgrey-registers` crate, which allows
writing registers to a peripheral which is borrowed as read-only.
But the semantics of i.e. reading or writing from the hardware UART
involves modifying the state of the peripheral.

Marking the UART methods as `&mut self` allows the borrow checker to
help enforce this.