Experiment with a standalone Cargo build.
Cargo normally allows minimal customization to invoke the linker.
Notably, this limited by default to allowing passing additional
libraries and library directories when invoking the linker.

However, it is possible to specify a custom linker command in the
Cargo configuration for a build triplet. The only limitation is that
this is a single string, specifying a path to invoke. This prevents
directly using this value to pass custom link commands, such as the
linker script to use.

However, it is possible to use a simple shim script to do our dirty
work. The shim invokes the linker with the necessary custom options,
and appends the arguments passed by `rustc`.

The result is a fully linked executable.

Downsides?

The name is chosen by Cargo, and doesn't have a `.elf` extension. The
default behavior of Verilator is to use the extension to determine if
it's an ELF or VMEM file. The workaround is to inform Verilator of the
file type by appending ",elf" to the `--meminit` argument.

The executable build this way is slightly larger, with an additional
24 bytes in the .text segment. I believe this is caused by rustc
forcing `main` to have the (argc, argv) -> exit-code signature. The
symbol table is also slightly bigger, as it now includes `main()`.

This also doesn't solve how the linker shim gets created in the first
place, however `build.rs` can likely provide the necessary hooks to
create it.