Use custom target and  `cargo-xbuild` to build.
This creates a custom rustc target config to specify the linker and
apply some customization to the build.

This isn't a slam-dunk approach. Even though the triplet matched the
existing `riscv32imc-unknown-elf` triplet, Cargo still attempts to
compile against a renamed core library (the name appears to contain
the hash of the file?). This in turn requires `cargo-xbuild` to
provide the core dependency.

And while it allows customizing both the linker executable and any
custom linker arguments (a plus over just a custom .cargo/config) it
still requires some trickery to link in the CRT assembly.

Given that this still requires additional setup (`cargo install cargo
xbuild`) and still doesn't solve cutting out GCC, it doesn't seem an
ideal solution.